requirements.txt
cv2
imutils
numpy
pyzbar.pyzbar
json
uwsgi